modulo-gestao-estrategica/servico-acompanhamento-indicadores/app/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
requests==2.27.1
python-dateutil==2.8.2
